SDL3 GUI step 7: GUI M1 — board rendering, input staging, message log
bin/oracle-gui is now genuinely playable: wired to SessionClient, renders
the real board every frame, accepts mouse input, and logs what happened.
Verified via a full human-vs-AI playtest (mulligan-era ATTACK, champion
toggle+confirm, Decline, draw-card follow-up all exercised live) plus
repeated valgrind passes (0 errors, 0 leaked bytes throughout).

- stda_gui.c/gui_app.c: pre-window stdio player setup (mirrors stda_tui.c),
  session_local_start() wiring, SDL_SetRenderVSync (was pegging a full CPU
  core — SDL_AppIterate has no built-in frame cap).
- gui_layout.c: region rects from window size, indexed by seat slot (not
  raw PlayerID) so a future 3-4 player rework only touches this file.
- gui_palette.c: card border colours from the HSL thematic-colours table,
  cross-checked against the printed card sheets by rasterizing and
  sampling actual pixels rather than trusting the table alone.
- gui_card.c/gui_render.c: procedural (no-art-yet) card rendering — border
  colour, species/cost/dice/base-attack text — plus the status bar, action
  bar, and per-seat info, all through LOCALIZED_STRING_L (EN/FR/ES).
- gui_input.c: click-to-stage input per the synthesis doc's §9.4 table
  (champion toggle, draw-vs-recall branch with a discard overlay, cash
  exchange targeting, Pass/Decline). Confirm's enabled state mirrors
  SessionUpdate.legal[] as a UX convenience only — engine_submit()'s
  decision_is_legal() remains the real gate. Verified via a throwaway
  integration harness driving real coordinates from gui_layout.c against
  a hand-built gamestate (not committed — mouse automation isn't available
  in this environment).
- gui_log.c: scrolling event log (moves, combat with real numbers already
  in CombatDetails, luna, discards, game over), fed from
  SessionUpdate.events.

Two ComicNeue-Regular/PatrickHand-Regular fonts (previous commit) render
status/card text respectively.

Also fixes a real bug gui_log.c surfaced: EventBuf (game_event.h) capped
at 32 events and silently dropped the rest (event_buf_push()), sized for
one engine step — but stda_session.c's publish cadence (Step 6) batches
an entire silent AI-only stretch into one EventBuf, which for a full
AI-vs-AI game is the whole game. Measured empirically (200 seeded
Borealis-vs-Daredevil games via a scratch harness, not committed): max
189 events in one game (39 turns, ~4.85/turn), mean 70.4. Bumped
EVENT_BUF_CAP to 512 (>2.7x headroom) and widened EventBuf.count from
uint8_t to uint16_t to actually support a cap above 255 — every uint8_t
loop variable scanning it was widened to match (game_event.c,
stda_session.c, gui_log.c, testsrc/test_game_engine.c), since a uint8_t
loop counter bounded by a uint16_t count would itself wrap at 255.

make test_game_event (15/15), make test_game_engine (28/28), make
test_session (8/8) all pass; ./bin/oracle -a -p matches
bin/expectedresults.txt byte-for-byte; make format clean.

+++ src/roles/stda/stda_session.c
@@ -56,7 +56,7 @@ static void fill_legal(SessionUpdate* u, const struct gamestate* gs, PendingDeci
 } // fill_legal
 
 static void publish(SessionClient* c, PendingDecision pending, EventBuf events, bool rejected)
-{ for(uint8_t i = 0; i < events.count; i++)
+{ for(uint16_t i = 0; i < events.count; i++)
     event_filter_for_viewer(&events.ev[i], c->viewer);
 
   SessionUpdate u = {0};
